refactor(types): reorganize candidate role types and related imports
- removed `Candidate`, `FormFields`, `AuthFormProps`, `TermsModalProps`, and `TermsData` exports from `types/index.ts`

- moved:
  - `Candidate` interface → `types/models/candidate.ts`
  - `FormFields`, `AuthFormProps`, `TermsModalProps`, `TermsData` → `types/auth/candidate/`

- updated import paths in:
  - favorite types and API files
  - `termsData` file
  - `useFavoriteCompany` custom hook
  - `CandidateForm` and `TermsModal` components

- in `api/firebase/candidate-signup` route, changed `createdWith` value from "Email/Password Provider" to "Form"
- in `api/favorites` route, added type for `item` parameter in `isAlreadyFavorited` function
