+++ lib/firebase.ts
@@ -26,9 +26,9 @@ export const SignInWithGooglePopup = async () => {
         displayName: credential.user.displayName,
         acceptedTerms: `${credential.user.displayName} adlı Kullanıcı, kullanım şartlarını kabul ettiğini ve bu şartlara uygun hareket edeceğini beyan eder.`,
         email: credential.user.email,
-        candidatePhoto: credential.user.photoURL,
+        photo: credential.user.photoURL,
         emailVerified: credential.user.emailVerified,
-        createdWith: "Google Provider",
+        createdWith: "Google",
         role: "candidate",
       };
